@@ -1,26 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.Extensions.Configuration;
+using System.IO;
 
 namespace CoreCodeCamp.Data
 {
-  public class CampContextFactory : IDesignTimeDbContextFactory<CampContext>
-  {
-    public CampContext CreateDbContext(string[] args)
+    public class CampContextFactory : IDesignTimeDbContextFactory<CampContext>
     {
-      var config = new ConfigurationBuilder()
-        .SetBasePath(Directory.GetCurrentDirectory())
-        .AddJsonFile("appsettings.json")
-        .Build();
+        public CampContext CreateDbContext(string[] args)
+        {
+            var config = new ConfigurationBuilder()
+              .SetBasePath(Directory.GetCurrentDirectory())
+              .AddJsonFile("appsettings.json")
+              .Build();
 
-      return new CampContext(new DbContextOptionsBuilder<CampContext>().Options, config);
+            return new CampContext(new DbContextOptionsBuilder<CampContext>().Options, config);
+        }
     }
-  }
 }
