@@ -1,16 +1,15 @@
 ﻿namespace CoreCodeCamp.Data
 {
-  public class Speaker
-  {
-    public int SpeakerId { get; set; }
-    public string FirstName { get; set; }
-    public string LastName { get; set; }
-    public string MiddleName { get; set; }
-    public string Company { get; set; }
-    public string CompanyUrl { get; set; }
-    public string BlogUrl { get; set; }
-    public string Twitter { get; set; }
-    public string GitHub { get; set; }
-
-  }
+    public class Speaker
+    {
+        public int SpeakerId { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string MiddleName { get; set; }
+        public string Company { get; set; }
+        public string CompanyUrl { get; set; }
+        public string BlogUrl { get; set; }
+        public string Twitter { get; set; }
+        public string GitHub { get; set; }
+    }
 }
