@@ -1,5 +1,4 @@
-﻿
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
@@ -12,30 +11,34 @@ namespace CoreCodeCamp.Models
         [Required]
         [StringLength(100)]
         public string Name { get; set; }
+
         // Serve de chave primária visível para o cliente.
         [Required]
-        public string Moniker { get; set; } 
+        public string Moniker { get; set; }
+
         public DateTime EventDate { get; set; } = DateTime.MinValue;
-        [Range(1,100)]
+
+        [Range(1, 100)]
         public int Length { get; set; } = 1;
 
         // Location
+        public LocationModel Location { get; set; }
         /*
          * Ao se utilizar o nome da entidade como prefixo,
-         * o AutoMapper irá realizar um bind automático, 
+         * o AutoMapper irá realizar um bind automático,
          * não sendo necessário criar uma LocationModel caso não queiramos retornar
          * todos os dados de Location associados a um camp
          */
-        public string Venue { get; set; }
-        public string LocationAddress1 { get; set; }
-        public string LocationAddress2 { get; set; }
-        public string LocationAddress3 { get; set; }
-        public string LocationCityTown { get; set; }
-        public string LocationStateProvince { get; set; }
-        public string LocationPostalCode { get; set; }
-        public string LocationCountry { get; set; }
+        //public string Venue { get; set; }
+        //public string LocationAddress1 { get; set; }
+        //public string LocationAddress2 { get; set; }
+        //public string LocationAddress3 { get; set; }
+        //public string LocationCityTown { get; set; }
+        //public string LocationStateProvince { get; set; }
+        //public string LocationPostalCode { get; set; }
+        //public string LocationCountry { get; set; }
 
-        //Talk     
+        //Talk
         public ICollection<TalkModel> Talks { get; set; }
     }
 }
