@@ -1,8 +1,12 @@
-﻿namespace CoreCodeCamp.Data
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace CoreCodeCamp.Models
 {
-    public class Location
+    public class LocationModel
     {
-        public int LocationId { get; set; }
         public string VenueName { get; set; }
         public string Address1 { get; set; }
         public string Address2 { get; set; }
